+++ internal/bot/events/guildcreate.go
@@ -8,20 +8,33 @@ import (
 	"time"
 )
 
-func OnGuildCreate(s *discordgo.Session, g *discordgo.GuildCreate) {
+// OnGuildCreate is called when the bot joins a new guild
+// This function gets called on two occassions: on startup and when the bot joins a new guild
+// We use this event to send a welcome message to the guild if we join a new guild.
+// We also use this event to count the number of members in each voice channel
+func OnGuildCreate(session *discordgo.Session, guild *discordgo.GuildCreate) {
 
-	slog.Debug("Guild create event", "guild", g.Guild.ID, "name", g.Guild.Name)
+	slog.Debug("Guild create event", "guild", guild.Guild.ID, "name", guild.Guild.Name)
 
-	// Count members in each voice channel
-	for _, vs := range g.VoiceStates {
+	setOccupancy(guild.Guild)
+
+	if time.Since(guild.JoinedAt) < time.Minute {
+		notifications.SendWelcomeMessage(session, guild.Guild)
+	}
+}
+
+// setOccupancy sets the occupancy for the guild
+func setOccupancy(guild *discordgo.Guild) {
+	occpuancy := 0
+
+	for _, vs := range guild.VoiceStates {
 		if vs.ChannelID != "" {
-			storage.IncrementOccupancy(g.ID)
+			occpuancy += 1
 		}
 	}
 
-	slog.Info("Setting the occupancy for the guild", "guild", g.Guild.ID, "occupancy", storage.GetOccupancy(g.Guild.ID))
+	storage.SetOccupancy(guild.ID, occpuancy)
 
-	if time.Since(g.JoinedAt) < time.Minute {
-		notifications.SendWelcomeMessage(s, g.Guild)
-	}
+	slog.Info("Setting the occupancy for the guild", "guild", guild.ID, "occupancy", occpuancy)
+
 }