+++ internal/bot/events/interactioncreate.go
@@ -6,11 +6,20 @@ import (
 	"log/slog"
 )
 
-func OnInteractionCreate(s *discordgo.Session, i *discordgo.InteractionCreate) {
-	slog.Debug("Interaction create event", "interaction", i.ID, "type", i.Type, "guild", i.GuildID, "channel", i.ChannelID, "member", i.Member.User.ID)
+// OnInteractionCreate is called when an interaction is created
+// We use this to handle slash commands
+func OnInteractionCreate(session *discordgo.Session, interaction *discordgo.InteractionCreate) {
+	slog.Debug(
+		"Interaction create event",
+		"interaction", interaction.ID,
+		"type", interaction.Type,
+		"guild", interaction.GuildID,
+		"channel", interaction.ChannelID,
+		"member", interaction.Member.User.ID,
+	)
 
-	switch i.Type {
+	switch interaction.Type {
 	case discordgo.InteractionApplicationCommand:
-		commands.HandleCommand(s, i)
+		commands.HandleCommand(session, interaction)
 	}
 }