+++ internal/commands/enable.go
@@ -6,32 +6,34 @@ import (
 	"log/slog"
 )
 
+// EnableCommand is the command to enable notifications
 var enableCommand = &discordgo.ApplicationCommandOption{
 	Name:        "enable",
 	Type:        discordgo.ApplicationCommandOptionSubCommand,
 	Description: "Enable notifications when people start voice chatting in this server",
 }
 
-func handleEnableCommand(s *discordgo.Session, i *discordgo.InteractionCreate) {
-	guild := i.GuildID
+// handleEnableCommand handles the enable command
+func handleEnableCommand(session *discordgo.Session, interaction *discordgo.InteractionCreate) {
+	guild := interaction.GuildID
 
 	var err error
-	message := "As soon as someone starts voice chatting in this server, you will be notified"
+	message := ":white_check_mark: As soon as someone starts voice chatting in this server, I will send you a message!"
 
-	if i.Member == nil {
-		message = "You can only use this command in a server"
+	if interaction.Member == nil {
+		message = ":exclamation: You can only use this command in a server"
 	} else {
-		err = storage.RegisterUser(i.Member.User.ID, guild)
+		err = storage.RegisterUser(interaction.Member.User.ID, guild)
 
 		if err != nil {
-			slog.Error("Could not register user", "user", i.Member.User.ID, "guild", guild, "error", err)
-			message = "Could not register you. Please try again later."
+			slog.Error("Could not register user", "user", interaction.Member.User.ID, "guild", guild, "error", err)
+			message = ":exclamation: Could not register you. Please try again later."
 		} else {
-			slog.Info("User registered for guild", "user", i.Member.User.ID, "guild", guild)
+			slog.Info("User registered for guild", "user", interaction.Member.User.ID, "guild", guild)
 		}
 	}
 
-	err = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+	err = session.InteractionRespond(interaction.Interaction, &discordgo.InteractionResponse{
 		Type: discordgo.InteractionResponseChannelMessageWithSource,
 		Data: &discordgo.InteractionResponseData{
 			Content: message,