+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Styled the messages that are sent to users
 
+### Fixed
+- Fixed a potential concurrency issue
 ## [0.1.0] - 2024-12-10
 ### Added
 - Added initial version of the bot